refactor: do not define on init file
